requirements.txt
torch
transformers
streamlit
Pillow
accelerate
timm